Fixed issue with Requare parameter. Added function for RecreateOrder

+++ src/LykkePayAPI/Models/Markup.cs
@@ -5,11 +5,11 @@ namespace LykkePay.API.Models
 {
     public class Markup
     {
-        public float Percent { get; set; }
+        public double Percent { get; set; }
 
         public int Pips { get; set; }
 
-        public float FixedFee { get; set; }
+        public double FixedFee { get; set; }
 
         
     }
